requirements.txt
fastapi
uvicorn
opencv-python
requests
tqdm
Pillow
numpy
python-dotenv
